// src/lib/types.ts
export interface Category {
  _id: string;
  name: string;
  slug: string;
  description: string;
  image: string;
  isActive: boolean;
}

export interface CustomizationChoice {
  value: string;
  priceModifier: number;
}

export interface CustomizationOption {
  _id: string;
  type: "material" | "color" | "size" | "engraving" | "add-on" | "other";
  label: string;
  required: boolean;
  choices: CustomizationChoice[];
}

export interface ProductImage {
  url: string;
  publicId: string;
}

export interface Product {
  _id: string;
  name: string;
  slug: string;
  description: string;
  category: Category | string;
  basePrice: number;
  images: ProductImage[];
  customizationOptions: CustomizationOption[];
  stock: number;
  lowStockThreshold: number;
  isFeatured: boolean;
  isActive: boolean;
  isCustomizable: boolean;
  createdAt: string;
  updatedAt: string;
}